+++ apps/web/src/layout/AppMenu.vue
@@ -19,9 +19,10 @@ const model = ref([
                 icon: 'pi pi-fw pi-server',
                 items: [
                     { label: 'Siswa', icon: 'pi pi-fw pi-bars', to: '/presences' },
+                    { label: 'Statistik', icon: 'pi pi-fw pi-chart-line', to: '/stats' },
                 ]
             },
-            { label: 'Statistik', icon: 'pi pi-fw pi-chart-line', to: '/stats' },
+            { label: 'Payment', icon: 'pi pi-fw pi-wallet', to: '/payment' },
             { label: 'Profil', icon: 'pi pi-fw pi-user', to: '/profile' },
             { label: 'Backup', icon: 'pi pi-fw pi-database', to: '/backup' },
             // {