+++ apps/web/src/views/Payment/All.vue
@@ -1,0 +1,3 @@
+<template>
+    <p>Comming Soon</p>
+</template>